Move AlphaFilter out of EKF
