+++ AlphaFilter/AlphaFilter.hpp
